Refactor the AlpineJS for Menus and Modals
Removed the duplication of toggle and close functions and put together the states of the menus. Also, similar refactoring of the modal. Lastly added sm:hidden to backdrop so that when the sidemenu is active from a screen size lower than sm to a larger screen size the backdrop won't stay in the way.